requirements.txt
# 核心依赖
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0
httpx>=0.25.2
openai>=1.3.0
PyJWT>=2.8.0

# 飞书机器人依赖
lark-oapi>=1.3.0
requests>=2.31.0

# 工具依赖
python-dotenv>=1.0.0
PyYAML>=6.0.0

# 开发依赖
# pytest>=7.4.0
# black>=23.11.0
# isort>=5.12.0
